requirements.txt
asgiref==3.11.0
braintree==4.41.0
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap3==2024.1
django==5.2.9
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.2
django-registration-redux==2.13
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
pillow==12.1.0
pyjwt==2.10.1
requests==2.32.5
sqlparse==0.5.5
typing-extensions==4.15.0
urllib3==2.6.2